separate firebase from sagas
