+++ src/firebase/events.js
@@ -1,6 +1,6 @@
 import firebase from 'firebase/app'
 
-import crud from 'sagas/firebase/crud'
+import crud from './crud'
 
 export const fetchUserEvents = uid =>
   firebase
