+++ src/firebase/user.js
@@ -1,0 +1,3 @@
+import crud from './crud'
+
+export default crud('users', 'uid')
