--- src/sagas/auth/index.js
@@ -1,1 +1,0 @@
-export { default } from './auth.saga'
