--- src/sagas/events/index.js
@@ -1,1 +1,0 @@
-export { default } from './events.saga'
