--- src/sagas/firebase/index.js
@@ -1,1 +1,0 @@
-export { default } from './firebase.saga'
