+++ src/sagas/index.js
@@ -1,14 +1,14 @@
 import { all } from 'redux-saga/effects'
 
-import firebaseSaga from './firebase'
-import authSaga from './auth'
-import eventsSaga from './events'
-import userSaga from './user'
-import talksSaga from './talks'
-import speakerSaga from './speaker'
-import submissionSaga from './submission'
-import proposalsSaga from './proposals'
-import ratingsSaga from './ratings'
+import firebaseSaga from './firebase.saga'
+import authSaga from './auth.saga'
+import eventsSaga from './events.saga'
+import userSaga from './user.saga'
+import talksSaga from './talks.saga'
+import speakerSaga from './speaker.saga'
+import submissionSaga from './submission.saga'
+import proposalsSaga from './proposals.saga'
+import ratingsSaga from './ratings.saga'
 
 export default function* sagas() {
   try {
