--- src/sagas/proposals/index.js
@@ -1,1 +1,0 @@
-export { default } from './proposals.saga'
