+++ src/sagas/ratings.saga.js
@@ -3,8 +3,8 @@ import { call, select, takeLatest } from 'redux-saga/effects'
 import store from 'redux/store'
 import { getRatingsAverage } from 'redux/data/ratings.selectors'
 import { getRouterParam } from 'redux/router'
-import { getRatings, addRating } from './ratings.firebase'
-import { updateRating } from '../proposals/proposals.firebase'
+import { getRatings, addRating } from 'firebase/ratings'
+import { updateRating } from 'firebase/proposals'
 
 function* fetchRatings({ eventId, proposalId }) {
   // wipe current ratings
