--- src/sagas/ratings/index.js
@@ -1,1 +1,0 @@
-export { default } from './ratings.saga'
