--- src/sagas/submission/index.js
@@ -1,1 +1,0 @@
-export { default } from './submission.saga'
