--- src/sagas/talks/index.js
@@ -1,1 +1,0 @@
-export { default } from './talks.saga'
