--- src/sagas/user/index.js
@@ -1,1 +1,0 @@
-export { default } from './user.saga'
