--- src/sagas/user/user.firebase.js
@@ -1,3 +1,0 @@
-import crud from 'sagas/firebase/crud'
-
-export default crud('users', 'uid')
